+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - New function `remove_member` to remove a member from a list
   - New function `subscribe` to subscribe to a list
   - New function `unsubscribe` to unsubscribe from a list
+  - New function `add_member_list` to add multiple users to a list
 - New enum `service::ListMethod` and `list` field in `service::RateLimitStatus` to contain the rate
   limit info for these methods